UniqueLottoNumbersValidatable 로 고유한 번호 검증 로직 분리
